+++ Ambassador/src/utopia/ambassador/database/factory/scope/ScopeFactory.scala
@@ -1,0 +1,21 @@
+package utopia.ambassador.database.factory.scope
+
+import utopia.ambassador.database.AmbassadorTables
+import utopia.ambassador.model.partial.scope.ScopeData
+import utopia.ambassador.model.stored.scope.Scope
+import utopia.flow.datastructure.immutable.{Constant, Model}
+import utopia.flow.generic.ValueUnwraps._
+import utopia.vault.nosql.factory.FromValidatedRowModelFactory
+
+/**
+  * Used for reading scope data from the DB
+  * @author Mikko Hilpinen
+  * @since 11.7.2021, v1.0
+  */
+object ScopeFactory extends FromValidatedRowModelFactory[Scope]
+{
+	override def table = AmbassadorTables.scope
+	
+	override protected def fromValidatedModel(model: Model[Constant]) = Scope(model("id"),
+		ScopeData(model("serviceId"), model("serviceSideName"), model("clientSideName")))
+}